+++ tests/testthat/test-extract.R
@@ -190,6 +190,8 @@ test_that("error if not a workflow", {
 # extract_parameter_set_dials()
 
 test_that("extract parameter set from workflow with tunable recipe", {
+  skip_if(tune_contains_tunable_method, message = "`tunable.model_spec()` is in tune")
+
   spline_rec <- recipes::recipe(ridership ~ ., data = head(Chicago)) %>%
     recipes::step_date(date) %>%
     recipes::step_holiday(date) %>%
@@ -211,6 +213,8 @@ test_that("extract parameter set from workflow with tunable recipe", {
 })
 
 test_that("extract parameter set from workflow with tunable model", {
+  skip_if(tune_contains_tunable_method, message = "`tunable.model_spec()` is in tune")
+
   rm_rec <- recipes::recipe(ridership ~ ., data = head(Chicago)) %>%
     recipes::step_rm(date, ends_with("away"))
   bst_model <-
@@ -225,6 +229,8 @@ test_that("extract parameter set from workflow with tunable model", {
 })
 
 test_that("extract parameter set from workflow with tunable recipe and model", {
+  skip_if(tune_contains_tunable_method, message = "`tunable.model_spec()` is in tune")
+
   spline_rec <- recipes::recipe(ridership ~ ., data = head(Chicago)) %>%
     recipes::step_date(date) %>%
     recipes::step_holiday(date) %>%
@@ -254,6 +260,8 @@ test_that("extract parameter set from workflow with tunable recipe and model", {
 # extract_parameter_dials()
 
 test_that("extract single parameter from workflow with tunable recipe", {
+  skip_if(tune_contains_tunable_method, message = "`tunable.model_spec()` is in tune")
+
   spline_rec <- recipes::recipe(ridership ~ ., data = head(Chicago)) %>%
     recipes::step_date(date) %>%
     recipes::step_holiday(date) %>%
@@ -288,6 +296,8 @@ test_that("extract single parameter from workflow with tunable recipe", {
 })
 
 test_that("extract single parameter from workflow with tunable model", {
+  skip_if(tune_contains_tunable_method, message = "`tunable.model_spec()` is in tune")
+
   rm_rec <- recipes::recipe(ridership ~ ., data = head(Chicago)) %>%
     recipes::step_rm(date, ends_with("away"))
   bst_model <-
@@ -306,6 +316,8 @@ test_that("extract single parameter from workflow with tunable model", {
 })
 
 test_that("extract single parameter from workflow with tunable recipe and model", {
+  skip_if(tune_contains_tunable_method, message = "`tunable.model_spec()` is in tune")
+
   spline_rec <- recipes::recipe(ridership ~ ., data = head(Chicago)) %>%
     recipes::step_date(date) %>%
     recipes::step_holiday(date) %>%